Address review: use RenderRoamBlock util and extract isSmartBlockUid
- Use RenderRoamBlock from ~/utils/roamReactComponents (correct API
  path: ui.react.Block, not ui.components.react.block). Fixes runtime
  break and removes the @ts-expect-error + eslint-disables.
- Move isSmartBlockUid into its own file at ~/utils/isSmartBlockUid.ts.
