+++ apps/roam/src/utils/getLeftSidebarSettings.ts
@@ -8,13 +8,6 @@ import {
   getUidAndStringSetting,
 } from "./getExportSettings";
 import { getSubTree } from "roamjs-components/util";
-import getTextByBlockUid from "roamjs-components/queries/getTextByBlockUid";
-
-export const isSmartBlockUid = (uid: string): boolean => {
-  const text = getTextByBlockUid(uid);
-  if (!text) return false;
-  return text.includes(":SmartBlock:");
-};
 
 type LeftSidebarPersonalSectionSettings = {
   uid: string;
